+++ rnd/autogpt_server/autogpt_server/server/__init__.py
@@ -1,3 +1,3 @@
-from .server import AgentServer
+from .api_server import AgentServer
 
 __all__ = ["AgentServer"]
