+++ rnd/autogpt_server/autogpt_server/util/test.py
@@ -5,7 +5,7 @@ from autogpt_server.data.block import Block, initialize_blocks
 from autogpt_server.data.execution import ExecutionStatus
 from autogpt_server.executor import ExecutionManager, ExecutionScheduler
 from autogpt_server.server import AgentServer
-from autogpt_server.server.server import get_user_id
+from autogpt_server.server.api_server import get_user_id
 from autogpt_server.util.service import PyroNameServer
 
 log = print
